chore: rollback

+++ pago-backend/src/main/java/tw/pago/pagobackend/service/impl/OtpServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import tw.pago.pagobackend.dao.OtpDao;
 import tw.pago.pagobackend.dto.SmsRequestDto;
 import tw.pago.pagobackend.dto.ValidatePhoneRequestDto;
+import tw.pago.pagobackend.exception.TooManyRequestsException;
 import tw.pago.pagobackend.model.Otp;
 import tw.pago.pagobackend.service.OtpService;
 import tw.pago.pagobackend.service.SmsService;
@@ -42,12 +43,12 @@ public class OtpServiceImpl implements OtpService {
         
             long secondsRemaining = cooldownInSeconds - differenceInSeconds;
 
-            // TODO REMOVE THIS AFTER TESTING
-//            if (differenceInSeconds < cooldownInSeconds) {
-//                System.out.println("It's been less than 3 minutes since the last reset request");
-//              // It's been less than 3 minutes since the last reset request
-//              throw new TooManyRequestsException("You can request another SNS in " + secondsRemaining + " seconds.", latestResetDateTime, secondsRemaining);
-//            }
+
+            if (differenceInSeconds < cooldownInSeconds) {
+                System.out.println("It's been less than 3 minutes since the last reset request");
+              // It's been less than 3 minutes since the last reset request
+              throw new TooManyRequestsException("You can request another SNS in " + secondsRemaining + " seconds.", latestResetDateTime, secondsRemaining);
+            }
             otpDao.deleteOtpById(existingOtp.getOtpId());
         }
